requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
appdirs==1.4.4
    # via fs
arrow==1.3.0
    # via cookiecutter
asgiref==3.8.1
    # via django
backports-zoneinfo==0.2.1 ; python_version < "3.9"
    # via
    #   -c requirements/constraints.txt
    #   django
binaryornot==0.4.4
    # via cookiecutter
boto3==1.34.79
    # via fs-s3fs
botocore==1.34.79
    # via
    #   boto3
    #   s3transfer
build==1.2.1
    # via pip-tools
certifi==2024.2.2
    # via requests
chardet==5.2.0
    # via
    #   binaryornot
    #   diff-cover
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   cookiecutter
    #   pip-tools
cookiecutter==2.6.0
    # via xblock-sdk
coverage[toml]==7.4.4
    # via pytest-cov
ddt==1.7.2
    # via -r requirements/test.in
diff-cover==8.0.3
    # via -r requirements/dev.in
    # via
    #   -c requirements/common_constraints.txt
    #   -r requirements/base.in
    #   edx-i18n-tools
    #   openedx-django-pyfs
    #   xblock-sdk
edx-i18n-tools==1.3.0
    # via -r requirements/test.in
exceptiongroup==1.2.0
    # via pytest
fs==2.4.16
    # via
    #   fs-s3fs
    #   openedx-django-pyfs
    #   xblock
fs-s3fs==1.1.1
    # via
    #   openedx-django-pyfs
    #   xblock-sdk
idna==3.6
    # via requests
importlib-metadata==6.11.0
    # via
    #   -c requirements/common_constraints.txt
    #   build
iniconfig==2.0.0
    # via pytest
jinja2==3.1.3
    # via
    #   cookiecutter
    #   diff-cover
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lazy==1.6
    # via xblock
lxml==4.9.4
    # via
    #   edx-i18n-tools
    #   xblock
    #   xblock-sdk
mako==1.3.2
    # via xblock
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
    #   xblock
mdurl==0.1.2
    # via markdown-it-py
mock==5.1.0
    # via -r requirements/test.in
nose==1.3.7
    # via -r requirements/test.in
openedx-django-pyfs==3.6.0
    # via xblock
packaging==24.0
    # via
    #   build
    #   pytest
path==16.13.0
    # via edx-i18n-tools
pip-tools==7.4.1
    # via -r requirements/dev.in
pluggy==1.4.0
    # via
    #   diff-cover
    #   pytest
polib==1.2.0
    # via edx-i18n-tools
pygments==2.17.2
    # via
    #   diff-cover
    #   rich
pypng==0.20220715.0
    # via
    #   -r requirements/dev.in
    #   xblock-sdk
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
pytest==8.1.1
    # via
    #   -r requirements/test.in
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r requirements/test.in
pytest-django==4.8.0
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   botocore
    #   xblock
python-slugify==8.0.4
    # via cookiecutter
pytz==2024.1
    # via xblock
pyyaml==6.0.1
    # via
    #   cookiecutter
    #   edx-i18n-tools
    #   xblock
requests==2.31.0
    # via
    #   -r requirements/base.in
    #   cookiecutter
    #   xblock-sdk
rich==13.7.1
    # via cookiecutter
s3transfer==0.10.1
    # via boto3
simplejson==3.19.2
    # via
    #   xblock
    #   xblock-sdk
six==1.16.0
    # via
    #   fs
    #   fs-s3fs
    #   python-dateutil
sqlparse==0.4.4
    # via django
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   build
    #   coverage
    #   pip-tools
    #   pyproject-hooks
    #   pytest
types-python-dateutil==2.9.0.20240316
    # via arrow
typing-extensions==4.11.0
    # via
    #   asgiref
    #   rich
urllib3==1.26.18
    # via
    #   botocore
    #   requests
web-fragments==2.2.0
    # via
    #   xblock
    #   xblock-sdk
webob==1.8.7
    # via
    #   xblock
    #   xblock-sdk
wheel==0.43.0
    # via pip-tools
xblock[django]==3.1.0
    # via
    #   -r requirements/base.in
    #   xblock-sdk
xblock-sdk==0.10.0
    # via
    #   -r requirements/dev.in
    #   -r requirements/test.in
zipp==3.18.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
